CRAFTERCMS-1094 Submit to delete services implemented
