+++ src/main/webapp/default-site/static-assets/components/cstudio-dialogs-templates/submit-for-delete.html
@@ -40,7 +40,7 @@
     </div>
 
     <div class="action-wrapper">
-        <button class="btn btn-primary schedule-for-delete">Submit</button>
+        <button class="btn btn-primary submit-for-delete">Submit</button>
         <button class="btn btn-default cancel">Cancel</button>
     </div>
